+++ build_runner/lib/src/entrypoint/test.dart
@@ -81,7 +81,6 @@ class TestCommand extends BuildRunnerCommand {
         builderApplications,
         deleteFilesByDefault: options.deleteFilesByDefault,
         enableLowResourcesMode: options.enableLowResourcesMode,
-        failOnSevere: options.failOnSevere,
         configKey: options.configKey,
         assumeTty: options.assumeTty,
         outputMap: outputMap,